rating/requirements.txt
#Python=3.11
Flask==2.1
requests==2.28.2
opentelemetry-sdk
opentelemetry-api
opentelemetry-instrumentation-flask==0.36b0
opentelemetry-instrumentation-requests==0.36b0
opentelemetry-exporter-otlp
waitress==2.1.2